MEXSVR-247　[KINOP] resTransactionでの未処理D/Cのチェック SQLからcount(*)で取得する値はintで運用する
git-svn-id: http://svnsin.ncr.com/svn/repos/pa/res/RES%20MeX/Base/resTransaction/branches/3.3.0@67645 9234da4d-b7fe-4b44-acf3-10f21a033f33

+++ test/ncr/res/mobilepos/poslogstatus/test/PoslogStatusTestSteps.java
@@ -65,13 +65,13 @@ public class PoslogStatusTestSteps extends Steps {
 	}
 	
 	@Then("a consolidationResult is $consolidationResult")
-	public final void theConsolidationResultIs(final Long consolidationResult) {
+	public final void theConsolidationResultIs(final int consolidationResult) {
 		Assert.assertEquals("Expect the  consolidationResult code", consolidationResult,
 				poslogStatusInfo.getConsolidationResult());
 	}
 	
 	@Then("a transferResult is $transferResult")
-	public final void theTransferResultIs(final Long transferResult) {
+	public final void theTransferResultIs(final int transferResult) {
 		Assert.assertEquals("Expect the  transferResult code", transferResult,
 				poslogStatusInfo.getTransferResult());
 	}
